refactor(route): add BusStopDrawer
