+++ apps/bus/src/pages/city/[citySlug]/[routeName].tsx
@@ -2,15 +2,9 @@ import React, { useEffect, useState } from 'react';
 
 import {
   Box,
-  Button,
   Circle,
-  Drawer,
-  DrawerBody,
-  DrawerContent,
-  DrawerHeader,
   Flex,
   Heading,
-  HStack,
   IconButton,
   Stack,
   Tab,
@@ -45,16 +39,13 @@ import type {
 } from 'next';
 import { useRouter } from 'next/router';
 import NextHeadSeo from 'next-head-seo';
-import { BiChevronLeft, BiChevronRight } from 'react-icons/bi';
+import { BiChevronLeft } from 'react-icons/bi';
 import { BsInfoCircle } from 'react-icons/bs';
 import { IoHome } from 'react-icons/io5';
-import { MdClose } from 'react-icons/md';
 import { GeoJSONLineString, parse } from 'wellknown';
 
-import background from '@/background-small.png';
-import bus from '@/bus.png';
 import BusRouteInfoModal from '@/components/BusRouteInfoModal';
-import Image from '@/components/Image';
+import BusStopDrawer, { ZoomLevel } from '@/components/BusStopDrawer';
 import { DESKTOP_MAP_LEFT } from '@/components/Layout';
 import { useMap } from '@/components/MapContextProvider';
 import NavBarItems from '@/components/NavBarItems';
@@ -63,7 +54,7 @@ import {
   busEstimationSelector,
   useGetBusEstimationQuery,
 } from '@/services/local';
-import { getLastElement, getMiddleElement } from '@/utils/array';
+import { getMiddleElement } from '@/utils/array';
 import { getTwoDigitString } from '@/utils/string';
 
 interface BusRoutePageProps {
@@ -73,13 +64,6 @@ interface BusRoutePageProps {
   route: BusRouteDetail;
   directions: BusDirection[];
   routeStopEntity: RouteStopEntity;
-}
-
-enum ZoomLevel {
-  Stop = 16,
-  Stops = 15,
-  Marker = 13.5,
-  City = 12,
 }
 
 type RouteStopEntity = Record<BusDirection, RouteStop>;
@@ -143,34 +127,6 @@ const BusRoutePage = ({
 
   const onHomeClick = () => {
     router.push('/');
-  };
-
-  const handleRouteStopClick = async (step: 1 | -1) => {
-    const currentStops = selectedBusRoute.Stops;
-
-    const previousStop =
-      currentStops[
-        currentStops.findIndex((stop) => stop.StopUID === selectedStopId) + step
-      ];
-
-    setSelectedStopId(previousStop.StopUID);
-
-    const { getPosition } = await import('@/services/mapbox');
-    mapContextRef.current.map.flyTo(
-      getPosition(
-        previousStop.StopPosition.PositionLat,
-        previousStop.StopPosition.PositionLon,
-        ZoomLevel.Stop,
-      ),
-    );
-  };
-
-  const onPreviousStopClick = async () => {
-    await handleRouteStopClick(-1);
-  };
-
-  const onNextStopClick = async () => {
-    await handleRouteStopClick(1);
   };
 
   useEffect(() => {
@@ -494,102 +450,15 @@ const BusRoutePage = ({
       </Flex>
       <BusRouteInfoModal isOpen={isOpen} onClose={onClose} route={route} />
       {selectedStop && (
-        <Drawer
-          isOpen={stopDisclosure.isOpen}
+        <BusStopDrawer
+          selectedStopId={selectedStopId}
+          setSelectedStopId={setSelectedStopId}
+          route={route}
+          selectedStop={selectedStop}
+          selectedBusRoute={selectedBusRoute}
           onClose={onDrawerClose}
-          size="lg"
-          closeOnOverlayClick={false}
-          placement="bottom"
-        >
-          <DrawerContent minH="200px" textAlign="center">
-            <Box
-              pos="fixed"
-              w="full"
-              h="full"
-              overflow="hidden"
-              bg="gradient.bg"
-              zIndex="1"
-            >
-              <Image
-                alt="background"
-                src={background}
-                placeholder="blur"
-                layout="fill"
-                objectFit="cover"
-                objectPosition="bottom"
-              />
-            </Box>
-            <DrawerHeader pb="0" zIndex="docked" noOfLines={1}>
-              {selectedStop.StopName.Zh_tw}
-            </DrawerHeader>
-            <IconButton
-              pos="absolute"
-              right="4"
-              top="5"
-              rounded="full"
-              size="xs"
-              color="primary.600"
-              bgColor="primary.50"
-              aria-label="close modal"
-              fontSize="xl"
-              zIndex="docked"
-              icon={<MdClose />}
-              onClick={onDrawerClose}
-            />
-            <DrawerBody display="flex" flexDir="column" pt="0" zIndex="docked">
-              <Text noOfLines={1} color="primary.200">
-                往
-                {selectedStop.Direction === BusDirection.去程
-                  ? route.DestinationStopNameZh
-                  : route.DepartureStopNameZh}
-              </Text>
-              <HStack mx="auto" my="2">
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  leftIcon={<BiChevronLeft />}
-                  isDisabled={
-                    selectedBusRoute.Stops[0].StopUID === selectedStopId
-                  }
-                  onClick={onPreviousStopClick}
-                >
-                  上一站
-                </Button>
-                <Text minW="40px">
-                  {/* TODO: add stop status util to better support different scenario */}
-                  {selectedStop?.StopStatus === BusStopStatus.正常
-                    ? `${
-                        selectedStop?.EstimateTime
-                          ? `${Math.floor(selectedStop?.EstimateTime / 60)}分`
-                          : '進站中'
-                      }`
-                    : '今日未營運'}{' '}
-                </Text>
-                <Button
-                  variant="ghost"
-                  rightIcon={<BiChevronRight />}
-                  size="sm"
-                  isDisabled={
-                    getLastElement(selectedBusRoute.Stops).StopUID ===
-                    selectedStopId
-                  }
-                  onClick={onNextStopClick}
-                >
-                  下一站
-                </Button>
-              </HStack>
-              <Box pos="absolute" bottom="1" left="0" right="0">
-                <Image
-                  src={bus}
-                  placeholder="blur"
-                  width={200}
-                  height={60}
-                  objectFit="contain"
-                />
-              </Box>
-            </DrawerBody>
-          </DrawerContent>
-        </Drawer>
+          isOpen={stopDisclosure.isOpen}
+        />
       )}
     </>
   );
